+++ tests/configs_tests.adb
@@ -16,12 +16,14 @@ package body Configs_Tests is
                                       Arg  ("root",     'r', "Content root (default: ""content"" in the current directory)"),
                                       Arg  ("cert",     'c', "TLS certificate path (default: ""cert.pem"" in the current directory)"),
                                       Arg  ("key",      'k', "TLS key path (default: ""key.pem"" in the current directory)"),
+                                      Arg  ("workers",  'w', "Workers count (default: 8)"),
                                       Flag ("help",     'h', "Print this message")];
 
    procedure Test_Valid_Arguments (T : in out Test_Context) is
       Arguments : constant Opts.Argument_List :=
          ["-H", "example.com", "-p", "1024", "-r", "resources/gemini/content",
-          "-c", "resources/gemini/cert.pem", "-k", "resources/gemini/key.pem"];
+          "-c", "resources/gemini/cert.pem", "-k", "resources/gemini/key.pem",
+          "-w", "4"];
       Result : constant Opts.Result := Opts.Parse (Arguments, Options);
 
       Actual : constant Configs.Config := Configs.Parse (Result);
@@ -31,6 +33,7 @@ package body Configs_Tests is
       T.Expect (Actual.Content_Root = "resources/gemini/content", "Content_Root: expected: resources/gemini/content, got: " & Actual.Content_Root);
       T.Expect (Actual.Cert_File = "resources/gemini/cert.pem", "Cert_File: expected: resources/gemini/cert.pem, got: " & Actual.Cert_File);
       T.Expect (Actual.Key_File = "resources/gemini/key.pem", "Key_File: expected: resources/gemini/key.pem, got: " & Actual.Key_File);
+      T.Expect (Actual.Workers_Count = 4, "Workers_Count: expected: 4, got:" & Actual.Workers_Count'Image);
    end Test_Valid_Arguments;
 
    procedure Invalid_Port is
@@ -77,7 +80,7 @@ package body Configs_Tests is
 
    procedure Invalid_Key_File is
       Arguments : constant Argument_List := ["--root", "resources/gemini/content", "--cert", "resources/gemini/cert.pem",
-                                              "--key", "foo.pem"];
+                                             "--key", "foo.pem"];
       Result : constant Opts.Result := Opts.Parse (Arguments, Options);
 
       Unused_Config : Configs.Config;
@@ -89,4 +92,19 @@ package body Configs_Tests is
    begin
       T.Expect_Raises (Invalid_Key_File'Access, Configs.Config_Error'Identity, "invalid key file: foo.pem");
    end Test_Invalid_Key_File;
+
+   procedure Invalid_Workers_Count is
+      Arguments : constant Argument_List := ["--root", "resources/gemini/content", "--cert", "resources/gemini/cert.pem",
+                                             "--key", "resources/gemini/key.pem", "-w", "foo"];
+      Result : constant Opts.Result := Opts.Parse (Arguments, Options);
+
+      Unused_Config : Configs.Config;
+   begin
+      Unused_Config := Configs.Parse (Result);
+   end Invalid_Workers_Count;
+
+   procedure Test_Invalid_Workers_Count (T : in out Test_Context) is
+   begin
+      T.Expect_Raises (Invalid_Workers_Count'Access, Configs.Config_Error'Identity, "invalid workers count: foo");
+   end Test_Invalid_Workers_Count;
 end Configs_Tests;