+++ tests/configs_tests.ads
@@ -8,4 +8,5 @@ package Configs_Tests is
    procedure Test_Invalid_Root (T : in out Test_Context);
    procedure Test_Invalid_Cert_File (T : in out Test_Context);
    procedure Test_Invalid_Key_File (T : in out Test_Context);
+   procedure Test_Invalid_Workers_Count (T : in out Test_Context);
 end Configs_Tests;